Move noteGraph scaffolding to utils
