+++ packages/foam-core/test/janitor/generateLinkReferences.test.ts
@@ -1,12 +1,13 @@
+import * as path from 'path';
 import { NoteGraph, Note } from '../../src/note-graph';
 import { generateLinkReferences } from '../../src/janitor';
-import { scaffold } from '../__scaffold__';
+import { initializeNoteGraph } from '../../src/utils';
 
 describe('generateLinkReferences', () => {
   let _graph: NoteGraph;
 
   beforeAll(async () => {
-    _graph = await scaffold();
+    _graph = await initializeNoteGraph(path.join(__dirname, '../__scaffold__'));
   });
 
   it('initialised test graph correctly', () => {
@@ -43,7 +44,7 @@ describe('generateLinkReferences', () => {
 
   it('should remove link definitions from a file that has them, if no links are present', () => {
     const note = _graph.getNote('second-document') as Note;
-    
+
     const expected = {
       newText: "",
       range: {
@@ -59,7 +60,7 @@ describe('generateLinkReferences', () => {
         },
       },
     };
-    
+
     const actual = generateLinkReferences(note!, _graph);
 
     expect(actual!.range.start).toEqual(expected.range.start);
@@ -69,7 +70,7 @@ describe('generateLinkReferences', () => {
 
   it('should update link definitions if they are present but changed', () => {
     const note = _graph.getNote('first-document') as Note;
-    
+
     const expected = {
       newText: `[file-without-title]: file-without-title "file-without-title"`,
       range: {
@@ -85,7 +86,7 @@ describe('generateLinkReferences', () => {
         },
       },
     };
-    
+
     const actual = generateLinkReferences(note!, _graph);
 
     expect(actual!.range.start).toEqual(expected.range.start);
@@ -95,9 +96,9 @@ describe('generateLinkReferences', () => {
 
   it('should not cause any changes if link reference definitions were up to date', () => {
     const note = _graph.getNote('third-document') as Note;
-    
+
     const expected = null;
-    
+
     const actual = generateLinkReferences(note!, _graph);
 
     expect(actual).toEqual(expected);
